feat: Add in-memory routing
FEATURE:  Memory-only routing
Fixes #7.

It is a delta LOC of +117 and I'm not fully convinced that this brings value.

What would be the use case for in-memory routing, other than "quick tests" or unit testing?

Still pending:  Unit testing.

Will let consumers of the package make a case for this feature.